+++ cmd/repo-updater/repoupdater/server.go
@@ -26,6 +26,9 @@ type Server struct {
 	repos.Store
 	*repos.Syncer
 	*repos.OtherReposSyncer
+	InternalAPI interface {
+		ReposUpdateMetadata(ctx context.Context, repo api.RepoName, description string, fork, archived bool) error
+	}
 }
 
 // Handler returns the http.Handler that should be used to serve requests.
@@ -220,7 +223,7 @@ func (s *Server) repoLookup(ctx context.Context, args protocol.RepoLookupArgs) (
 		}
 		if repo != nil {
 			go func() {
-				err := api.InternalClient.ReposUpdateMetadata(context.Background(), repo.Name, repo.Description, repo.Fork, repo.Archived)
+				err := s.InternalAPI.ReposUpdateMetadata(context.Background(), repo.Name, repo.Description, repo.Fork, repo.Archived)
 				if err != nil {
 					log15.Warn("Error updating repo metadata", "repo", repo.Name, "err", err)
 				}