+++ enterprise/cmd/frontend/db/external_services_test.go
@@ -259,6 +259,55 @@ func TestExternalServices_ValidateConfig(t *testing.T) {
 			assert: excludes(`authorization.ttl: time: invalid duration 0`),
 		},
 		{
+			kind:   "GITHUB",
+			desc:   "invalid empty exclude",
+			config: `{"exclude": []}`,
+			assert: includes(`exclude: Array must have at least 1 items`),
+		},
+		{
+			kind:   "GITHUB",
+			desc:   "invalid empty exclude item",
+			config: `{"exclude": [{}]}`,
+			assert: includes(`exclude.0: Must validate at least one schema (anyOf)`),
+		},
+		{
+			kind:   "GITHUB",
+			desc:   "invalid exclude item",
+			config: `{"exclude": [{"foo": "bar"}]}`,
+			assert: includes(`exclude.0: Must validate at least one schema (anyOf)`),
+		},
+		{
+			kind:   "GITHUB",
+			desc:   "invalid exclude item name",
+			config: `{"exclude": [{"name": "bar"}]}`,
+			assert: includes(`exclude.0.name: Does not match pattern '^[\w-]+/[\w.-]+$'`),
+		},
+		{
+			kind:   "GITHUB",
+			desc:   "invalid empty exclude item id",
+			config: `{"exclude": [{"id": ""}]}`,
+			assert: includes(`exclude.0.id: String length must be greater than or equal to 1`),
+		},
+		{
+			kind:   "GITHUB",
+			desc:   "invalid additional exclude item properties",
+			config: `{"exclude": [{"id": "foo", "bar": "baz"}]}`,
+			assert: includes(`bar: Additional property bar is not allowed`),
+		},
+		{
+			kind: "GITHUB",
+			desc: "both name and id can be specified in exclude",
+			config: `
+			{
+				"url": "https://github.corp.com",
+				"token": "very-secret-token",
+				"exclude": [
+					{"name": "foo/bar", "id": "AAAAA="}
+				]
+			}`,
+			assert: equals(`<nil>`),
+		},
+		{
 			kind:   "GITLAB",
 			desc:   "without url nor token",
 			config: `{}`,